Update for new automation syntax
The automation example for notification of the certificate didn't renew did not work when using configuration.yaml. I added an example that do work with the new format. Not sure if the old example should still be kept.

Followed the rest of the guide and it works like a charm (still have to wait 60 days to test the renewal though)